requirement.txt
absl-py==1.4.0
anyio==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.2
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
boltons==23.0.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.4
cmake==3.25.0
comm==0.1.3
conda==23.5.0
conda-content-trust==0.1.3
conda-package-handling==1.9.0
contourpy==1.1.0
cryptography==39.0.1
cycler==0.11.0
daal==2023.1.1
daal4py==2023.1.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
einops==0.8.1
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.17.1
filelock==3.9.0
fonttools==4.40.0
fqdn==1.5.1
google-auth==2.21.0
google-auth-oauthlib==1.0.0
grpcio==1.56.0
idna==3.4
imageio==2.35.1
importlib-metadata==6.7.0
importlib-resources==5.12.0
ipykernel==6.23.3
ipython==8.12.3
ipython-genutils==0.2.0
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
json5==0.9.14
jsonpatch==1.32
jsonpointer==2.1
jsonschema==4.17.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter-events==0.6.3
jupyter-lsp==2.2.0
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.2
jupyterlab-language-pack-zh-CN==4.0.post0
jupyterlab-pygments==0.2.2
jupyterlab_server==2.23.0
keras==2.13.1
kiwisolver==1.4.4
lit==15.0.7
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==3.0.1
mpmath==1.2.1
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.16.6
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==3.0
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.24.1
nvidia-ml-py==11.525.131
nvitop==1.1.2
oauthlib==3.2.2
overrides==7.3.1
packaging==23.0
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pip==23.1.2
pipreqs==0.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.8.0
pluggy==1.0.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycosat==0.6.4
pycparser==2.21
Pygments==2.15.1
pyOpenSSL==23.0.0
pyparsing==3.1.0
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
requests==2.29.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.3.0
scikit-learn-intelex==2023.1.1
scipy==1.10.1
Send2Trash==1.8.2
setuptools==67.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
sympy==1.11.1
tbb==2021.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.1.1
tomli==2.0.1
toolz==0.12.0
torch==2.0.0+cu118
torchaudio==2.0.1+cu118
torchvision==0.15.1+cu118
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typing_extensions==4.4.0
tzdata==2023.3
uri-template==1.3.0
urllib3==1.25.8
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.3.6
wheel==0.38.4
xgboost==1.7.6
yarg==0.1.9
zipp==3.15.0
